+++ backend/src/controllers/csvController.js
@@ -1,34 +1,8 @@
 import csvtojson from 'csvtojson';
-import fs from 'fs';
-import path from 'path';
 import { Student } from '../models/student.model.js';
 import { Semester } from '../models/sem.model.js';
-import { Subject } from '../models/subjects.model.js';
+import createStudentFolders from './folderController.js';
 
-// Function to create student folders
-const subjects = Subject;
-const createStudentFolders = (student, subjects) => {
-    const baseDir = path.join('.', 'public', 'uploads', student.rollno);
-    const assignmentDir = path.join(baseDir, 'assignments');
-    const materialsDir = path.join(baseDir, 'study-materials');
-
-    // Create base directory
-    fs.mkdirSync(baseDir, { recursive: true });
-    
-    // Create subject folders inside assignments
-    subjects.forEach(subject => {
-        fs.mkdirSync(path.join(assignmentDir, subject.name), { recursive: true });
-    });
-
-    // Create the study materials folder
-    fs.mkdirSync(materialsDir, { recursive: true });
-
-    // Return the paths (optional)
-    return {
-        assignmentFolderPath: assignmentDir,
-        studyMaterialsFolderPath: materialsDir,
-    };
-};
 
 // Function to handle CSV import and folder creation
 const importUserCsv = async (req, res) => {
@@ -67,6 +41,7 @@ const importUserCsv = async (req, res) => {
         // Insert all student data into the database
         await Student.insertMany(userData);
         res.send({ status: 200, success: true, msg: 'CSV IMPORTED' });
+
     } catch (error) {
         res.status(400).send({ status: 400, success: false, msg: error.message });
     }
